+++ typed_sql/example/lib/src/model.g.dart
@@ -418,6 +418,8 @@ final class _$Package extends Package {
         columns: ['ownerId'],
         referencedTable: 'users',
         referencedColumns: ['userId'],
+        onDelete: .noAction,
+        onUpdate: .noAction,
       ),
     ],
     readRow: _$Package._$fromDatabase,