+++ typed_sql/test/typed_sql/foreign_key/nullable_composite_foreign_key/nullable_composite_foreign_key_test.g.dart
@@ -431,6 +431,8 @@ final class _$Book extends Book {
         columns: ['authorFirstName', 'authorLastName'],
         referencedTable: 'authors',
         referencedColumns: ['firstName', 'lastName'],
+        onDelete: .noAction,
+        onUpdate: .noAction,
       ),
     ],
     readRow: _$Book._$fromDatabase,