+++ typed_sql/test/typed_sql/foreign_key/simple_foreign_key/simple_foreign_key_test.g.dart
@@ -414,6 +414,8 @@ final class _$Book extends Book {
         columns: ['authorId'],
         referencedTable: 'authors',
         referencedColumns: ['authorId'],
+        onDelete: .noAction,
+        onUpdate: .noAction,
       ),
     ],
     readRow: _$Book._$fromDatabase,