+++ typed_sql/test/typed_sql/group_by/group_by_join_expression/group_by_join_expression_test.g.dart
@@ -322,6 +322,8 @@ final class _$Employee extends Employee {
         columns: ['deptId'],
         referencedTable: 'departments',
         referencedColumns: ['id'],
+        onDelete: .noAction,
+        onUpdate: .noAction,
       ),
     ],
     readRow: _$Employee._$fromDatabase,