+++ typed_sql/test/typed_sql/group_by/group_by_reference/group_by_reference_test.g.dart
@@ -368,6 +368,8 @@ final class _$Book extends Book {
         columns: ['authorId'],
         referencedTable: 'authors',
         referencedColumns: ['authorId'],
+        onDelete: .noAction,
+        onUpdate: .noAction,
       ),
     ],
     readRow: _$Book._$fromDatabase,