+++ typed_sql/test/typed_sql/join/join_using/join_using_test.g.dart
@@ -95,6 +95,8 @@ final class _$Employee extends Employee {
         columns: ['departmentId'],
         referencedTable: 'departments',
         referencedColumns: ['departmentId'],
+        onDelete: .noAction,
+        onUpdate: .noAction,
       ),
     ],
     readRow: _$Employee._$fromDatabase,